# Conflicts:
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktype/DailyWork.java
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/com/time/AttendanceClock.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/dailyperformanceformat/KrcmtBusinessTypeMonthlyPK.java
@@ -1,7 +1,6 @@
 package nts.uk.ctx.at.record.infra.entity.dailyperformanceformat;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
@@ -23,5 +22,5 @@ public class KrcmtBusinessTypeMonthlyPK implements Serializable{
 	public String businessTypeCode;
 	
 	@Column(name = "ATTENDANCE_ITEM_ID")
-	public BigDecimal attendanceItemId;
+	public int attendanceItemId;
 }
